Corrected use of subrouting and tests
